requirements.txt
gradio
neo4j
numpy
pandas
pillow
python-dotenv
opencv-python
matplotlib
chromadb
langchain
langgraph
langchain-openai
openai
transformers
torch
sentence-transformers
pure-python-adb
asyncio
aiohttp
lxml
langchain-google-genai